chore(showcase): #[expect] remaining C-parity structural lints + fix no-op casts/derefs
Tightest-scope expects (C-parity reasons) for collapsible-if/manual-clamp/
too-many-args/etc.; fn-level expects for assign_op_pattern + identity_op
sub-expressions where a statement attribute is rejected by stable Rust
(E0658); removed Rust-only no-op casts (unnecessary_cast) and derefs
(explicit_auto_deref) that have no C counterpart.

+++ showcase/examples/shaders/shaders_deferred_rendering.rs
@@ -486,11 +486,8 @@ fn main() {
 
                 // SAFETY: copy depth buffer from g-buffer to default framebuffer.
                 unsafe {
-                    ffi::rlBindFramebuffer(
-                        ffi::RL_READ_FRAMEBUFFER as u32,
-                        g_buffer.framebuffer_id,
-                    );
-                    ffi::rlBindFramebuffer(ffi::RL_DRAW_FRAMEBUFFER as u32, 0);
+                    ffi::rlBindFramebuffer(ffi::RL_READ_FRAMEBUFFER, g_buffer.framebuffer_id);
+                    ffi::rlBindFramebuffer(ffi::RL_DRAW_FRAMEBUFFER, 0);
                     ffi::rlBlitFramebuffer(
                         0,
                         0,
